+++ api/common/server/i18n.go
@@ -2,6 +2,7 @@ package server
 
 import (
 	"context"
+	"strings"
 
 	"github.com/gin-gonic/gin"
 	"golang.org/x/text/language"
@@ -9,11 +10,17 @@ import (
 	"dillmann.com.br/nginx-ignition/core/common/i18n"
 )
 
+const maximumLanguageTags = 10
+
 func i18nMiddleware(commands i18n.Commands) gin.HandlerFunc {
 	return func(ginCtx *gin.Context) {
 		lang := commands.DefaultLanguage()
 
 		langHeader := ginCtx.GetHeader("Accept-Language")
+		if strings.Count(langHeader, "-")+strings.Count(langHeader, "_") > maximumLanguageTags {
+			langHeader = ""
+		}
+
 		tags, _, err := language.ParseAcceptLanguage(langHeader)
 		if err == nil && len(tags) > 0 {
 			for _, tag := range tags {